+++ pkg/units/shell/k8s_manifest/factory.go
@@ -1,8 +1,8 @@
-package tfmodule
+package base
 
 import (
 	"github.com/apex/log"
-	"github.com/shalb/cluster.dev/pkg/modules/shell/terraform/base"
+	"github.com/shalb/cluster.dev/pkg/units/shell/common"
 	"github.com/shalb/cluster.dev/pkg/project"
 )
 
@@ -10,33 +10,37 @@ import (
 type Factory struct {
 }
 
-const unitKind string = "tfmodule"
+const unitKind string = "k8s-manifest"
 
-func NewEmptyUnit() Unit {
+func NewEmptyUnit() *Unit {
 	unit := Unit{
-		Unit:     *base.NewEmptyUnit(),
-		UnitKind: unitKind,
+		Unit:           *common.NewEmptyUnit(),
+		ManifestsFiles: &common.FilesListT{},
+		UnitKind:       unitKind,
+		ApplyTemplate:  true,
 	}
-	return unit
+	return &unit
 }
 
 func NewUnit(spec map[string]interface{}, stack *project.Stack) (*Unit, error) {
-	unit := NewEmptyUnit()
-	cUnit, err := base.NewUnit(spec, stack)
+	mod := NewEmptyUnit()
+
+	cUnit, err := common.NewUnit(spec, stack)
 	if err != nil {
 		log.Debug(err.Error())
 		return nil, err
 	}
-	unit.Unit = *cUnit
-	err = unit.ReadConfig(spec, stack)
+	mod.Unit = *cUnit
+	err = mod.ReadConfig(spec, stack)
 	if err != nil {
 		log.Debug(err.Error())
 		return nil, err
 	}
-	return &unit, nil
+	mod.BackendName = stack.BackendName
+	return mod, nil
 }
 
-// New creates new unit.
+// New creates new unit driver factory.
 func (f *Factory) New(spec map[string]interface{}, stack *project.Stack) (project.Unit, error) {
 	return NewUnit(spec, stack)
 }
@@ -44,18 +48,21 @@ func (f *Factory) New(spec map[string]interface{}, stack *project.Stack) (projec
 // NewFromState creates new unit from state data.
 func (f *Factory) NewFromState(spec map[string]interface{}, modKey string, p *project.StateProject) (project.Unit, error) {
 	mod := NewEmptyUnit()
+	// log.Fatal("FOOO")
 	err := mod.LoadState(spec, modKey, p)
 	if err != nil {
 		log.Debug(err.Error())
 		return nil, err
 	}
-	return &mod, nil
+	// modjs, _ := utils.JSONEncodeString(mod)
+	// log.Warnf("Mod from state: %v", modjs)
+	return mod, nil
 }
 
 func init() {
 	modDrv := Factory{}
 	log.Debugf("Registering unit driver '%v'", unitKind)
 	if err := project.RegisterUnitFactory(&modDrv, unitKind); err != nil {
-		log.Trace("Can't register unit driver '" + unitKind + "'.")
+		log.Fatalf("Can't register unit driver '%v'.", unitKind)
 	}
 }
