+++ pkg/units/shell/terraform/base/factory.go
@@ -1,30 +1,28 @@
-package helm
+package base
 
 import (
 	"github.com/apex/log"
-	"github.com/shalb/cluster.dev/pkg/modules/shell/terraform/base"
 	"github.com/shalb/cluster.dev/pkg/project"
+	"github.com/shalb/cluster.dev/pkg/units/shell/common"
 )
 
 // Factory factory for s3 backends.
 type Factory struct {
 }
 
-const unitKind string = "helm"
-
-func NewEmptyUnit() Unit {
+func NewEmptyUnit() *Unit {
 	unit := Unit{
-		Unit:     *base.NewEmptyUnit(),
-		HelmOpts: map[string]interface{}{},
-		Sets:     map[string]interface{}{},
-		UnitKind: unitKind,
+		Unit:              *common.NewEmptyUnit(),
+		RequiredProviders: make(map[string]RequiredProvider),
+		Initted:           false,
 	}
-	return unit
+	return &unit
 }
 
 func NewUnit(spec map[string]interface{}, stack *project.Stack) (*Unit, error) {
 	mod := NewEmptyUnit()
-	cUnit, err := base.NewUnit(spec, stack)
+
+	cUnit, err := common.NewUnit(spec, stack)
 	if err != nil {
 		log.Debug(err.Error())
 		return nil, err
@@ -35,7 +33,8 @@ func NewUnit(spec map[string]interface{}, stack *project.Stack) (*Unit, error) {
 		log.Debug(err.Error())
 		return nil, err
 	}
-	return &mod, nil
+	mod.BackendName = stack.BackendName
+	return mod, nil
 }
 
 // New creates new unit driver factory.
@@ -45,19 +44,13 @@ func (f *Factory) New(spec map[string]interface{}, stack *project.Stack) (projec
 
 // NewFromState creates new unit from state data.
 func (f *Factory) NewFromState(spec map[string]interface{}, modKey string, p *project.StateProject) (project.Unit, error) {
-	unit := NewEmptyUnit()
-	err := unit.LoadState(spec, modKey, p)
+	mod := NewEmptyUnit()
+	err := mod.LoadState(spec, modKey, p)
 	if err != nil {
 		log.Debug(err.Error())
 		return nil, err
 	}
-	return &unit, nil
-}
-
-func init() {
-	modDrv := Factory{}
-	log.Debugf("Registering unit driver '%v'", unitKind)
-	if err := project.RegisterUnitFactory(&modDrv, unitKind); err != nil {
-		log.Trace("Can't register unit driver '" + unitKind + "'.")
-	}
+	// modjs, _ := utils.JSONEncodeString(mod)
+	// log.Warnf("Mod from state: %v", modjs)
+	return mod, nil
 }
