+++ pkg/units/shell/terraform/helm/state.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/apex/log"
 
-	"github.com/shalb/cluster.dev/pkg/modules/shell/terraform/base"
 	"github.com/shalb/cluster.dev/pkg/project"
+	"github.com/shalb/cluster.dev/pkg/units/shell/terraform/base"
 	"github.com/shalb/cluster.dev/pkg/utils"
 )
