+++ pkg/units/shell/terraform/kubernetes/state.go
@@ -1,12 +1,12 @@
-package tfmodule
+package kubernetes
 
 import (
 	"fmt"
 
 	"github.com/apex/log"
 
-	"github.com/shalb/cluster.dev/pkg/modules/shell/terraform/base"
 	"github.com/shalb/cluster.dev/pkg/project"
+	"github.com/shalb/cluster.dev/pkg/units/shell/terraform/base"
 	"github.com/shalb/cluster.dev/pkg/utils"
 )
 
@@ -16,7 +16,6 @@ func (u *Unit) GetStateUnit() *Unit {
 	if err != nil {
 		log.Fatalf("read unit '%v': create state: %w", u.Name(), err)
 	}
-	unitState.OutputRaw = u.OutputRaw
 	unitState.Unit = *u.Unit.GetStateUnit()
 	return &unitState
 }
@@ -30,16 +29,16 @@ func (u *Unit) GetState() interface{} {
 
 type UnitDiffSpec struct {
 	base.UnitDiffSpec
-	Inputs        interface{} `json:"inputs"`
-	OutputsConfig *string     `json:"-"`
+	ProviderConf ProviderConfigSpec `json:"provider_conf"`
+	Inputs       interface{}        `json:"inputs"`
 }
 
 func (u *Unit) GetUnitDiff() UnitDiffSpec {
 	diff := u.Unit.GetUnitDiff()
 	st := UnitDiffSpec{
-		UnitDiffSpec:  diff,
-		Inputs:        u.Inputs,
-		OutputsConfig: nil,
+		UnitDiffSpec: diff,
+		ProviderConf: u.ProviderConf,
+		Inputs:       u.Inputs,
 	}
 	return st
 }
@@ -49,7 +48,6 @@ func (u *Unit) GetDiffData() interface{} {
 	res := map[string]interface{}{}
 	utils.JSONCopy(st, &res)
 	project.ScanMarkers(res, base.StringRemStScanner, u)
-	project.ScanMarkers(res, project.StateOutputsReplacer, u)
 	return res
 }
 
@@ -62,5 +60,5 @@ func (u *Unit) LoadState(stateData interface{}, modKey string, p *project.StateP
 	if err != nil {
 		return fmt.Errorf("load state: %v", err.Error())
 	}
-	return err
+	return nil
 }
