+++ pkg/units/shell/terraform/module/state.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/apex/log"
 
-	"github.com/shalb/cluster.dev/pkg/modules/shell/common"
-	"github.com/shalb/cluster.dev/pkg/modules/shell/terraform/base"
 	"github.com/shalb/cluster.dev/pkg/project"
+	"github.com/shalb/cluster.dev/pkg/units/shell/common"
+	"github.com/shalb/cluster.dev/pkg/units/shell/terraform/base"
 	"github.com/shalb/cluster.dev/pkg/utils"
 )
