+++ pkg/units/shell/terraform/printer/factory.go
@@ -1,40 +1,39 @@
-package base
+package tfmodule
 
 import (
 	"github.com/apex/log"
-	"github.com/shalb/cluster.dev/pkg/modules/shell/common"
 	"github.com/shalb/cluster.dev/pkg/project"
+	"github.com/shalb/cluster.dev/pkg/units/shell/terraform/base"
 )
 
 // Factory factory for s3 backends.
 type Factory struct {
 }
 
-func NewEmptyUnit() *Unit {
+const unitKind string = "printer"
+
+func NewEmptyUnit() Unit {
 	unit := Unit{
-		Unit:              *common.NewEmptyUnit(),
-		RequiredProviders: make(map[string]RequiredProvider),
-		Initted:           false,
+		Unit:     *base.NewEmptyUnit(),
+		UnitKind: unitKind,
 	}
-	return &unit
+	return unit
 }
 
 func NewUnit(spec map[string]interface{}, stack *project.Stack) (*Unit, error) {
-	mod := NewEmptyUnit()
-
-	cUnit, err := common.NewUnit(spec, stack)
+	unit := NewEmptyUnit()
+	cUnit, err := base.NewUnit(spec, stack)
 	if err != nil {
 		log.Debug(err.Error())
 		return nil, err
 	}
-	mod.Unit = *cUnit
-	err = mod.ReadConfig(spec, stack)
+	unit.Unit = *cUnit
+	err = unit.ReadConfig(spec, stack)
 	if err != nil {
 		log.Debug(err.Error())
 		return nil, err
 	}
-	mod.BackendName = stack.BackendName
-	return mod, nil
+	return &unit, nil
 }
 
 // New creates new unit driver factory.
@@ -44,13 +43,19 @@ func (f *Factory) New(spec map[string]interface{}, stack *project.Stack) (projec
 
 // NewFromState creates new unit from state data.
 func (f *Factory) NewFromState(spec map[string]interface{}, modKey string, p *project.StateProject) (project.Unit, error) {
-	mod := NewEmptyUnit()
-	err := mod.LoadState(spec, modKey, p)
+	unit := NewEmptyUnit()
+	err := unit.LoadState(spec, modKey, p)
 	if err != nil {
 		log.Debug(err.Error())
 		return nil, err
 	}
-	// modjs, _ := utils.JSONEncodeString(mod)
-	// log.Warnf("Mod from state: %v", modjs)
-	return mod, nil
+	return &unit, nil
+}
+
+func init() {
+	unitDrv := Factory{}
+	log.Debugf("Registering unit driver '%v'", unitKind)
+	if err := project.RegisterUnitFactory(&unitDrv, unitKind); err != nil {
+		log.Trace("Can't register unit driver '" + unitKind + "'.")
+	}
 }
