+++ pkg/units/shell/terraform/printer/state.go
@@ -1,12 +1,12 @@
-package kubernetes
+package tfmodule
 
 import (
 	"fmt"
 
 	"github.com/apex/log"
 
-	"github.com/shalb/cluster.dev/pkg/modules/shell/terraform/base"
 	"github.com/shalb/cluster.dev/pkg/project"
+	"github.com/shalb/cluster.dev/pkg/units/shell/terraform/base"
 	"github.com/shalb/cluster.dev/pkg/utils"
 )
 
@@ -16,6 +16,7 @@ func (u *Unit) GetStateUnit() *Unit {
 	if err != nil {
 		log.Fatalf("read unit '%v': create state: %w", u.Name(), err)
 	}
+	unitState.OutputRaw = u.OutputRaw
 	unitState.Unit = *u.Unit.GetStateUnit()
 	return &unitState
 }
@@ -29,16 +30,16 @@ func (u *Unit) GetState() interface{} {
 
 type UnitDiffSpec struct {
 	base.UnitDiffSpec
-	ProviderConf ProviderConfigSpec `json:"provider_conf"`
-	Inputs       interface{}        `json:"inputs"`
+	Inputs        interface{} `json:"inputs"`
+	OutputsConfig *string     `json:"-"`
 }
 
 func (u *Unit) GetUnitDiff() UnitDiffSpec {
 	diff := u.Unit.GetUnitDiff()
 	st := UnitDiffSpec{
-		UnitDiffSpec: diff,
-		ProviderConf: u.ProviderConf,
-		Inputs:       u.Inputs,
+		UnitDiffSpec:  diff,
+		Inputs:        u.Inputs,
+		OutputsConfig: nil,
 	}
 	return st
 }
@@ -48,6 +49,7 @@ func (u *Unit) GetDiffData() interface{} {
 	res := map[string]interface{}{}
 	utils.JSONCopy(st, &res)
 	project.ScanMarkers(res, base.StringRemStScanner, u)
+	project.ScanMarkers(res, project.StateOutputsReplacer, u)
 	return res
 }
 
@@ -60,5 +62,5 @@ func (u *Unit) LoadState(stateData interface{}, modKey string, p *project.StateP
 	if err != nil {
 		return fmt.Errorf("load state: %v", err.Error())
 	}
-	return nil
+	return err
 }
